refactor : handler를 controller와 컴포넌트트 render 함수 연결 역할로 분리
