+++ src/handler/resultHandler.js
@@ -1,8 +1,12 @@
-import { Modal } from "../components/Modal.js";
-import { Result } from "../components/Result.js";
-import { displayComponent } from "../util/displayComponents.js";
+import { ResultController } from "../controller/ResultController.js";
+import { WinningController } from "../controller/WinningController.js";
+import { calculateProfitRate } from "../service/ProfitService.js";
+import WebOutputView from "../view/WebOutputView.js";
 
-export const resultHandler = (matchingCount, profitRate) => {
-  const modalContent = Result({ matchingCount, profitRate });
-  displayComponent("#app", Modal({ content: modalContent }));
+export const resultHandler = async (lottoCount, lottoArray) => {
+  const winningLotto = await WinningController();
+  const matchingCount = ResultController(winningLotto, lottoArray);
+  const profitRate = calculateProfitRate(matchingCount, lottoCount);
+
+  WebOutputView.renderResult(matchingCount, profitRate);
 };
